Use Resttemplate and cron job pattern

+++ src/main/java/np/com/esewa/learn/sampleapplication/filedetails/repository/ProductFileRepository.java
@@ -8,7 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProductFileRepository extends JpaRepository<ProductFile, Long> {
-
+public interface ProductFileRepository extends JpaRepository<ProductFile, Integer> {
     List<ProductFile> findAllByStatus(FileStatus status);
 }
