+++ src/main/java/np/com/esewa/learn/sampleapplication/inventory/service/ProductService.java
@@ -1,6 +1,7 @@
 package np.com.esewa.learn.sampleapplication.inventory.service;
 
 import np.com.esewa.learn.sampleapplication.filedetails.dto.CountDto;
+import np.com.esewa.learn.sampleapplication.inventory.annotation.EncryptName;
 import np.com.esewa.learn.sampleapplication.inventory.dto.ProductDeleteDto;
 import np.com.esewa.learn.sampleapplication.inventory.dto.ProductResponseDto;
 import np.com.esewa.learn.sampleapplication.inventory.model.Product;
@@ -8,8 +9,9 @@ import np.com.esewa.learn.sampleapplication.inventory.model.Product;
 import java.util.List;
 
 public interface ProductService {
-    ProductResponseDto findById(Long id);
+    ProductResponseDto findById(int id);
+    /*ProductResponseDto findByCode(String code);*/
     List<Product> readProductDataFromFile(String filePath);
     CountDto addProduct(List<Product> productList);
-    ProductDeleteDto deleteProduct(Long id);
+    ProductDeleteDto deleteProduct(int id);
 }
